+++ packages/components/src/ActionList/ActionListHeader/index.ts
@@ -24,21 +24,6 @@
 
  */
 
-import { createContext } from 'react'
-import { MenuSharedProps } from './MenuItem'
-
-export interface MenuContextProps {
-  disabled?: boolean
-  id?: string
-  showDisclosure?: boolean
-  isOpen?: boolean
-  setOpen?: (isOpen: boolean) => void
-  triggerElement?: HTMLElement | null
-  triggerCallbackRef?: (node: HTMLElement | null) => void
-}
-
-const menuContext: MenuContextProps = {}
-const menuItemStyleContext: MenuSharedProps = {}
-
-export const MenuContext = createContext(menuContext)
-export const MenuItemStyleContext = createContext(menuItemStyleContext)
+export { ActionListHeader } from './ActionListHeader'
+export { ActionListHeaderColumn } from './ActionListHeaderColumn'
+export * from './generate'