+++ packages/components/src/ActionList/ActionListItemAction.tsx
@@ -24,21 +24,23 @@
 
  */
 
-import { createContext } from 'react'
-import { MenuSharedProps } from './MenuItem'
+import { CompatibleHTMLProps } from '@looker/design-tokens'
+import { IconNames } from '@looker/icons'
+import React, { ReactNode } from 'react'
+import { MenuItem } from '../Menu/'
 
-export interface MenuContextProps {
-  disabled?: boolean
-  id?: string
-  showDisclosure?: boolean
-  isOpen?: boolean
-  setOpen?: (isOpen: boolean) => void
-  triggerElement?: HTMLElement | null
-  triggerCallbackRef?: (node: HTMLElement | null) => void
+export interface ActionListItemActionProps
+  extends CompatibleHTMLProps<HTMLElement> {
+  children?: ReactNode
+  detail?: ReactNode
+  icon?: IconNames
+  role?: 'button' | 'link'
 }
 
-const menuContext: MenuContextProps = {}
-const menuItemStyleContext: MenuSharedProps = {}
-
-export const MenuContext = createContext(menuContext)
-export const MenuItemStyleContext = createContext(menuItemStyleContext)
+/**
+ * MenuItem may undergo a refactor soon. Creating a proxy in the form of ActionListItemAction
+ * allows us to adapt to any changes to MenuItem or its interface.
+ * */
+export const ActionListItemAction = (props: ActionListItemActionProps) => (
+  <MenuItem {...props} />
+)